requirements_production.txt
# ZAIN HMS Production Requirements - Optimized for Performance

# Core Django and Python packages
Django==5.2.6
python-decouple==3.8

# Database drivers - PostgreSQL optimized
psycopg2-binary==2.9.9
django-environ==0.11.2

# Caching and Performance
redis==5.0.1
django-redis==5.4.0
hiredis==2.2.3

# Background tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Web server and deployment
gunicorn==21.2.0
uvicorn[standard]==0.24.0
whitenoise==6.6.0

# Static files optimization
django-compressor==4.4
csscompressor==0.9.5
jsmin==3.0.1

# API and REST framework
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.26.5

# Authentication and Security
django-otp==1.3.0
qrcode==7.4.2
django-two-factor-auth==1.15.0

# File handling and uploads
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0  # For AWS S3 if needed

# Data processing and analytics
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Excel/CSV handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF generation
reportlab==4.0.7
weasyprint==60.2

# Email and communications
django-anymail==10.2

# Monitoring and debugging
sentry-sdk==1.40.0
django-debug-toolbar==4.2.0  # Development only
django-extensions==3.2.3

# Health checks
django-health-check==3.17.0

# Backup and maintenance
django-dbbackup==4.0.2

# Internationalization
django-rosetta==0.9.9

# Form handling
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
urllib3==2.1.0

# Development tools (remove in production)
django-silk==5.0.4  # Performance profiling
memory-profiler==0.61.0

# Testing (development only)
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.13.2

# Environment specific
python-dotenv==1.0.0